# contracts/openapi/core.v1.yaml
openapi: 3.1.0
info:
  title: ESWMP Core API
  version: v1
  description: |
    Scheduling primitives: Resources, Calendars, Availability Rules, Reservations,
    and Appointments. This service owns scheduling and nothing else — it never
    stores or returns any data belonging to the caller's own domain beyond the
    opaque externalReferenceType/externalReferenceId pair. See docs/ESWMP_VISION.md
    "Key rule".
servers:
  - url: http://localhost:6001
    description: Local development
security:
  - BearerAuth: []
paths:
...
  /api/v1/availability-rules:
    post:
      tags: [Availability]
      summary: Add a recurring availability rule for a Resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAvailabilityRuleRequest'
      responses:
        '201':
          description: Rule created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityRule'
  /api/v1/resources/{id}/availability:
    get:
      tags: [Availability]
      summary: Get a Resource's availability windows for a date
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: Availability windows for the date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityWindowsResponse'
  /api/v1/resources/{id}/time-off-requests:
    post:
      tags: [Availability]
      summary: Request time off for a Resource
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTimeOffRequest' }
      responses:
        '201': { description: Time off requested (PendingApproval), content: { application/json: { schema: { $ref: '#/components/schemas/TimeOff' } } } }
  /api/v1/time-off-requests/{id}/approve:
    post:
      tags: [Availability]
      summary: Approve a pending time off request
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Approved, content: { application/json: { schema: { $ref: '#/components/schemas/TimeOff' } } } }
        '409': { description: Not in PendingApproval }
  /api/v1/time-off-requests/{id}/reject:
    post:
      tags: [Availability]
      summary: Reject a pending time off request
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Rejected, content: { application/json: { schema: { $ref: '#/components/schemas/TimeOff' } } } }
        '409': { description: Not in PendingApproval }
  /api/v1/time-off-requests/{id}/cancel:
    post:
      tags: [Availability]
      summary: Cancel a pending or approved time off request
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Cancelled, content: { application/json: { schema: { $ref: '#/components/schemas/TimeOff' } } } }
        '409': { description: Not in PendingApproval or Approved }
  /api/v1/resources/{id}/availability-overrides:
    post:
      tags: [Availability]
      summary: Force-available or force-unavailable override for a Resource over a window
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAvailabilityOverrideRequest' }
      responses:
        '201': { description: Override created, content: { application/json: { schema: { $ref: '#/components/schemas/AvailabilityOverride' } } } }
  /api/v1/availability/resolve:
    post:
      tags: [Availability]
      summary: Resolve a Resource's free intervals over an arbitrary window, applying the full rule/exception/time-off/override priority stack
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResolveAvailabilityRequest' }
      responses:
        '200': { description: Free intervals, content: { application/json: { schema: { $ref: '#/components/schemas/ResolveAvailabilityResponse' } } } }
  /api/v1/availability/batch-resolve:
    post:
      tags: [Availability]
      summary: Resolve free intervals for several Resources at once
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BatchResolveAvailabilityRequest' }
      responses:
        '200':
          description: Free intervals per resource
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ResolveAvailabilityResponse' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateAvailabilityRuleRequest:
      type: object
      required: [resourceId, dayOfWeek, startTime, endTime]
      properties:
        resourceId: { type: string, format: uuid }
        dayOfWeek: { type: integer, minimum: 0, maximum: 6 }
        startTime: { type: string, example: "08:00" }
        endTime: { type: string, example: "18:00" }
        effectiveFrom: { type: string, format: date, nullable: true }
        effectiveTo: { type: string, format: date, nullable: true }
    AvailabilityRule:
      type: object
      required: [id, resourceId, dayOfWeek, startTime, endTime, ruleType, priority, status]
      properties:
        id: { type: string, format: uuid }
        resourceId: { type: string, format: uuid }
        availabilityProfileId: { type: string, format: uuid, nullable: true }
        dayOfWeek: { type: integer }
        startTime: { type: string }
        endTime: { type: string }
        effectiveFrom: { type: string, format: date, nullable: true }
        effectiveTo: { type: string, format: date, nullable: true }
        ruleType: { type: string, enum: [RegularWorkingHours, Shift, SeasonalSchedule, TemporarySchedule, AdditionalAvailability] }
        recurrencePattern:
          type: string
          nullable: true
          description: >-
            JSON-encoded { frequency: Daily|Weekly|SelectedWeekdays, interval, daysOfWeek[] }.
            Optional — when absent, dayOfWeek/startTime/endTime are authoritative.
        priority: { type: integer }
        status: { type: string, enum: [Active, Retired] }
        source: { type: string, nullable: true }
...
    CreateTimeOffRequest:
      type: object
      required: [type, startTime, endTime]
      properties:
        type: { type: string, example: "Vacation" }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        reasonCode: { type: string, nullable: true }
    TimeOff:
      type: object
      required: [id, resourceId, type, startTime, endTime, approvalStatus]
      properties:
        id: { type: string, format: uuid }
        resourceId: { type: string, format: uuid }
        type: { type: string }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        reasonCode: { type: string, nullable: true }
        approvalStatus: { type: string, enum: [Draft, PendingApproval, Approved, Rejected, Cancelled] }
    CreateAvailabilityOverrideRequest:
      type: object
      required: [effect, startTime, endTime]
      properties:
        effect: { type: string, enum: [ForceAvailable, ForceUnavailable] }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        reasonCode: { type: string, nullable: true }
    AvailabilityOverride:
      type: object
      required: [id, resourceId, effect, startTime, endTime, status]
      properties:
        id: { type: string, format: uuid }
        resourceId: { type: string, format: uuid }
        effect: { type: string, enum: [ForceAvailable, ForceUnavailable] }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        reasonCode: { type: string, nullable: true }
        status: { type: string, enum: [Active, Retired] }
    ResolveAvailabilityRequest:
      type: object
      required: [resourceId, startTime, endTime]
      properties:
        resourceId: { type: string, format: uuid }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
    ResolveAvailabilityResponse:
      type: object
      required: [resourceId, freeIntervals]
      properties:
        resourceId: { type: string, format: uuid }
        freeIntervals:
          type: array
          items:
            type: object
            properties:
              start: { type: string, format: date-time }
              end: { type: string, format: date-time }
    BatchResolveAvailabilityRequest:
      type: object
      required: [resourceIds, startTime, endTime]
      properties:
        resourceIds:
          type: array
          items: { type: string, format: uuid }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
...
    AvailabilityWindowsResponse:
      type: object
      required: [resourceId, date, windows]
      properties:
        resourceId: { type: string, format: uuid }
        date: { type: string, format: date }
        windows:
          type: array
          items:
            type: object
            properties:
              start: { type: string, format: date-time }
              end: { type: string, format: date-time }
...